refactor(frontend): extract session/tab core into sessionCore.ts (#312)
The heart of the SPA — the session list (refreshSessions / renderSessionList
/ openSession / updateToolbar) and the tab lifecycle (renderTabBar / openTab
/ addTab / closeTab) — moves out of main.ts into sessionCore.ts (~780 lines).

This is the "full split" territory: these functions reassign the
cross-cutting state (sessions, activeSessionId, currentTabs,
currentActiveTabId). ESM forbids reassigning imported bindings, so the state
stays declared in main.ts (exported live bindings) and main.ts gains setter
functions (setSessions / setActiveSessionId / setCurrentTabs /
setCurrentActiveTabId). sessionCore reads the live bindings directly and
writes through the setters — tsc enforces this (every reassignment was a
compile error until converted, so none can be missed). currentTerminals /
closingTabs (mutated, not reassigned) are plain exported consts.

Circular-import-safe: every main.ts import in sessionCore is used only
inside functions/handlers; no module-top access, so no TDZ. newSession.ts
repoints its openSession/refreshSessions/renderSessionList import to
sessionCore.

main.ts: 1766 → 1042 (and 5229 → 1042 over the whole #312 split — −80%).
tsc + Biome clean, vite build green, frontend tests pass (66).
